test: fixed test
